+++ apis/src/main/kotlin/org/yapp/apis/book/dto/response/BookSearchResponse.kt
@@ -1,7 +1,7 @@
 package org.yapp.apis.book.dto.response
 
 import org.yapp.domain.userbook.BookStatus
-import org.yapp.globalutils.util.IsbnConverter
+import org.yapp.apis.util.IsbnConverter
 import org.yapp.infra.external.aladin.response.AladinSearchResponse
 
 data class BookSearchResponse private constructor(
